Add qmllint hook with qt_precommit wrapper
The qt_precommit wrapper allows the Qt path to be configured by the
user. The wrapper will check for a `.qt_precommit` ini file, or use
the default `PATH` environment variable.

Provide a `qmllint` hook that uses the wrapper to help find the
`qmllint` tool, and sets `qmllint` to return a non-zero exit code if
there are issues. All other configuration should be provided through
the `.pre-commit-config.yaml` file.